数据库处理完成，现在不是事务操作，如果操作失败的话，应该是统一恢复到之前的值

+++ bookstore/src/com/zs/javaweb/domain/ShoppingCart.java
@@ -71,7 +71,7 @@ public class ShoppingCart {
 	}
 	
 	//计算购物车中的总额
-	public float getTotalMoney(){
+	public int getTotalMoney(){
 		int totalMoney = 0;
 		for(ShoppingCartItem item:books.values()){
 			totalMoney += item.getItemMoney();
